+++ packages/core/cairo_alpha/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-- Add ERC721Consecutive extension ([##769](https://github.com/OpenZeppelin/contracts-wizard/pull/#769))
+- Add ERC6909 ([#](https://github.com/OpenZeppelin/contracts-wizard/pull/))
+- Add ERC721 wrapper extension ([#764](https://github.com/OpenZeppelin/contracts-wizard/pull/764))
+- Add ERC721Consecutive extension ([#769](https://github.com/OpenZeppelin/contracts-wizard/pull/769))
 - Add ERC20 wrapper extension ([#763](https://github.com/OpenZeppelin/contracts-wizard/pull/763))
 
 - **Breaking changes**: